apply 42 norms and separate functions into multiple files
